fix: Fixed rb_put and rb_get functions.
rb_put and rb_get were previously created properly, but later modified to
fix a bug. The modification actually introduced a bug that this commit
fixes.

The functions were modified to increment the 'head' and 'tail' pointers
due to behavior at the serial terminal that seemed like it would have been
caused by the absence of a pointer increment in these functions. There was
a different cause of this behavior, and the functions were already properly
written, so the modification added a second, unnecessary increment.