feat: Added shut down message and user-LED toggle off.
When on-board button is pressed, the user-LED toggles on and an
initialization message prints to the terminal; this was not changed.
Once the the button has been toggled, it may now be pressed again to toggle
off, which will turn off the user-LED and print a shut down message to the
terminal.